Fix for ubuntu 10, apt-get update fails if dir doesn't exist (cherry picked from commit 8537806332028d1142a03e3380c884736f5ec1de)

+++ cookbooks/rightimage/libraries/util.rb
@@ -44,6 +44,10 @@ module RightImage
           FileUtils.rm_rf f         
         end
       end
+      # On 32 bit apt-get update fails if this directory doesn't exist
+      if ::File.directory? "/var/cache/apt/archives"
+        FileUtils.mkdir("/var/cache/apt/archives/partial", :mode=>0755)
+      end
 
       DIRS_truncate.each do |dir|
         files = ::Dir.glob(::File.join(@root, dir, "**", "*"))
